requirements.txt
pygame-ce
mediapipe
opencv-python
numpy